pztlibrary/requirements.txt
pytest~=8.3.5
pyserial~=3.5